Plasma: v0.2.1, correctness and engine fixes
- ESTIMATE: widen the weighted-Pauli term pool to u64 so terms on qubits
  16-27 no longer silently truncate to identity (was a u32 code).
- Stabilizer engine: ISWAP is now supported (composed as (S⊗S)×CZ×SWAP
  from the existing tableau updates); CSWAP/CPHASE stay excluded (they
  are genuinely non-Clifford).
- MPS engine: exact marginal sampling is now O(n×D^3) via a right
  environment plus a folded left density matrix, replacing the O(n^2×D^3)
  per-site bra-ket sandwich in the shot path.
- Remove the unused SimBackend::reset_backend trait method (no callers).
- Docs: correct the fusion/GPU description (the fused stream feeds only
  the --fuse CPU executor; the GPU uses its own mega-kernel encoding) and
  document the v0.2.1 behavior in PROJECT/README/SPEC.
- Tests: regression test for the ESTIMATE fix, a precise walk-vs-sandwich
  MPS marginals test, ISWAP coverage in the randomized stabilizer
  validation, and a fix for a test-thread stack overflow in
  test_clifford_detection. 99 tests + golden harness pass.

+++ src/gpu/mod.rs
@@ -1107,7 +1107,7 @@ impl Gpu {
                     let mut h = 0.0;
                     for k in off as usize..(off as usize + len as usize) {
                         let (c, p) = prog.estimate_terms[k];
-                        h += c * self.expect_value(p as u64)?;
+                        h += c * self.expect_value(p)?;
                     }
                     results.estimates.push(h);
                 }
